Move sessions fetching into SearchStore
